Factor out kernel symbolization logic into Ksyms type
Factor out the kernel symbolization logic into the newly created Ksyms
type. Doing so makes for cleaner separation of concerns and will make it
easier to swap out implementations in the future.
