+++ src/ksyms.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <cstdint>
+#include <string>
+
+class Ksyms {
+public:
+  Ksyms() = default;
+  ~Ksyms();
+
+  Ksyms(Ksyms &) = delete;
+  Ksyms &operator=(const Ksyms &) = delete;
+
+  std::string resolve(uint64_t addr, bool show_offset);
+
+private:
+  void *ksyms_{ nullptr };
+};
